requirements.txt
asgiref==3.8.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.3
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.9
django-ipware==7.0.1
django-session-timeout==0.1.0
idna==3.10
pillow==11.2.1
python-decouple==3.8
python-ipware==3.0.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
